+++ src/components/TodoBox.tsx
@@ -4,6 +4,8 @@ import TodoInput from './TodoInput';
 import TodoLists from './TodoLists';
 import styled from 'styled-components';
 import { Todo } from '../interface/Todo';
+import useAddTodo from '../hooks/useAddTodo';
+import useInput from '../hooks/useInput';
 
 const Title = styled.h2`
   margin: 0;
@@ -17,34 +19,8 @@ const ListTitle = styled.div`
   font-size: 17px;
 `;
 
-//localStorage key설정
-const TODOS_KEY = 'todos';
-
 const TodoBox = () => {
-  //localStorage에서 todos 가져오기
-  const [todos, setTodos] = useState(
-    JSON.parse(localStorage.getItem(TODOS_KEY)) || []
-  );
-
-  //localStorage에 todos 저장
-  useEffect(() => {
-    localStorage.setItem(TODOS_KEY, JSON.stringify(todos));
-  }, [todos]);
-
-  // input에 입력된 text로 todo 객체 생성
-  const handleTodoInput = useCallback(
-    (todoText: string) => {
-      const todo: Todo = {
-        id: Date.now(),
-        text: todoText,
-        isCompleted: false,
-      };
-
-      // 원래 있던 todo에 새로운 todo 붙이기
-      setTodos([...todos, todo]);
-    },
-    [todos]
-  );
+  const { todos, setTodos, handleTodoInput } = useAddTodo();
 
   // todo 삭제 function
   const handleTodoDelete = useCallback(
